refactor(was): User package 명 변경
