+++ src/main/java/entity/User.java
@@ -1,4 +1,4 @@
-package model;
+package entity;
 
 import java.util.Objects;
