+++ src/main/java/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package repository;
 
-import model.User;
+import entity.User;
 
 public interface UserRepository {
     User save(User user);
