+++ src/test/java/db/UserDataBaseTest.java
@@ -3,7 +3,7 @@ package db;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Collection;
-import model.User;
+import entity.User;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
